refactor: split runtime helpers out of const
Move runtime logic out of const.py into focused modules while
preserving all public behavior and backward-compatible re-exports.

New modules
-----------
registers/maps.py
  _build_map(), coil_registers(), discrete_input_registers(),
  holding_registers(), input_registers(), multi_register_sizes()

options/__init__.py
  OPTIONS_PATH, _load_json_option(), _get_options_init_lock(),
  async_setup_options(), SPECIAL_MODE_OPTIONS and all option list
  globals (DAYS_OF_WEEK, PERIODS, BYPASS_MODES, GWC_MODES,
  FILTER_TYPES, RESET_TYPES, MODBUS_PORTS, MODBUS_BAUD_RATES,
  MODBUS_PARITY, MODBUS_STOP_BITS)

entity_lookup.py
  _ENTITY_LOOKUP cache, _build_entity_lookup()

const.py after refactor
-----------------------
const.py now contains only constants, aliases, and thin façades.
All moved symbols are re-exported using the `X as X` pattern so
existing callers (including tests) continue to import from const
without changes.

_ENTITY_LOOKUP is intentionally kept in const so test suites can
inject fake lookups via monkeypatch.setattr(const_mod, "_ENTITY_LOOKUP",
fake); the const._build_entity_lookup() wrapper checks this attribute
first and delegates to entity_lookup._build_entity_lookup() when None.

Internal import updates
-----------------------
- switch.py, fan.py, climate.py: holding_/coil_registers from registers.maps
- entity.py: device_unique_id_prefix from unique_id_migration directly
- mappings/__init__.py: register map helpers from ..registers.maps
- _setup.py: async_setup_options from .options
- services_schema.py, config_flow.py: option globals from .options

No duplicate implementations; no circular imports; no behavior changes.
