+++ custom_components/thessla_green_modbus/mappings/__init__.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING, Any
 if TYPE_CHECKING:  # pragma: no cover
     from homeassistant.core import HomeAssistant
 
-from ..const import (
+from ..registers.maps import (
     coil_registers,
     discrete_input_registers,
     holding_registers,
