+++ custom_components/thessla_green_modbus/switch.py
@@ -13,11 +13,11 @@ from homeassistant.helpers.entity import EntityCategory
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .capability_rules import capability_block_reason
-from .const import coil_registers, holding_registers
 from .coordinator import ThesslaGreenModbusCoordinator
 from .entity import ThesslaGreenEntity
 from .mappings import ENTITY_MAPPINGS
 from .modbus_exceptions import ConnectionException, ModbusException
+from .registers.maps import coil_registers, holding_registers
 
 _LOGGER = logging.getLogger(__name__)
